Make it easier to add additional mount options
When using the `default` attribute of `mkOption`, the default value is
assigned with a very low priority and is overwritten by any other
assignment to the same option.
For the mount options, this will rarely be the desired behaviour, and
users will need to remember to add the default mount options themselves
when setting any additional mount options.

In this case, I think it's better to assign the default value with the
normal priority so that any additional assignments are simply added.
If a user really wants to overwrite the default options, they can use
`mkForce`.