improve ordering of mounts and tmpfiles
Previously tmpfiles was used to create the directory structure on both
the volatile and the persistent volume. Any bind-mounts from the former
to the latter were then set up afterwards.

This caused some issues, for example when preservation for `/var/log`
was set up too late, i.e. in stage 2, a dependency loop was created.
Prior to this change it was also impossible to use the default tmpfiles
services for managing files or directories inside a preserved directory
structure. The rationale was to use tmpfiles for the setup of missing
directory trees and files that should be preserved via bind-mounts,
both with correct permissions from the point of creation.

However, systemd mounts also support the creation of missing directories
(both source and target), but not the creation of bind-mounted source
files.

The new approach is as follows:
1. create missing source and target directory trees implicitly by
   bind-mounting preserved directories via systemd mounts
2. use the default tmpfiles service to create all missing files and
   symlinks + configure appropriate permissions for all preserved files
   and directories
3. create bind-mounts for files

This resolves the issue with the ordering cycle for `/var/log` and also
allows the use of the default tmpfiles service for managing files inside
directories that are preserved via bind-mount.

Note:
Using the nixos `fileSystems` module was considered but that does not
seem like a good fit:
- using the `fileSystems` module here is (currently) not possible:
  https://github.com/NixOS/nixpkgs/issues/24570
- it is obvious why mount units may be used for setting up bind-mounts,
  but configuring them via the `fileSystems` module the same way real
  fs mounts are configured, e.g. specifying `device`, is possible but
  quite far from intuitive and seems like an unnecessary indirection.